refactor: rename directory to yolo_swag_waste_grid_segmentation
