+++ models/yolo_swag_waste_grid_segmentation/Dockerfile.dockerignore
