+++ models/yolo_swag_waste_grid_segmentation/stac-item.json
@@ -45,12 +45,12 @@
   ],
   "properties": {
     "datetime": "2024-01-01T00:00:00Z",
-    "created": "2024-01-01T00:00:00Z",
-    "updated": "2024-01-01T00:00:00Z",
-    "title": "Open-access model for detecting openly dumped dispersed municipal solid waste from crowdsourced UAV imagery in Sub-Saharan Africa",
-    "description": "Managing openly dumped dispersed municipal solid waste in rapidly urbanizing Sub-Saharan Africa remains challenging due to dispersed informal dumping and limited high-resolution datasets for spatial monitoring. We present an open-access deep learning model for automated detection of openly dumped dispersed municipal solid waste via crowdsourced UAV imagery, trained and evaluated across 29 regions in 10 countries, encompassing diverse environmental contexts. A deep learning model trained on manually annotated 5 m × 5 m tiles achieved excellent performance in detecting openly dumped dispersed municipal solid waste across all study regions. Predicted distributions reveal heterogeneous waste accumulation patterns, ranging from localized hotspots—often along waterways, where waste can exacerbate flood and public health risks—to more dispersed litter across urban areas. Waste accumulation is most strongly associated with population density and indicators of lack of local infrastructure access, whereas its relationship with broader measures of regional development is weaker, highlighting the importance of fine-scale data for understanding localized waste dynamics. By releasing the model, this study provides a ready-to-use tool for UAV imagery collected by municipalities and local mapping communities, enabling openly dumped dispersed municipal solid waste monitoring without extensive technical expertise. This approach empowers local practitioners to convert UAV imagery into actionable insights, supporting targeted interventions and improved municipal solid waste management across Sub-Saharan Africa.",
-    "mlm:name": "yolo11xcls-waste-grid-segmentation",
-    "mlm:architecture": "yolo11xcls",
+    "created": "2026-07-20T00:00:00Z",
+    "updated": "2026-07-20T00:00:00Z",
+    "title": "YOLO Solid Waste Analysis on spatial Grids",
+    "description": "A image classification model for detecting solid waste piles from UAV imagery. Pretrained on 60 OAM scenes throughout the globe.",
+    "mlm:name": "yolo-swag-waste-grid-segmentation",
+    "mlm:architecture": "yolo26xcls",
     "mlm:tasks": [
       "semantic-segmentation"
     ],
@@ -74,7 +74,7 @@
           "host"
         ],
         "url": "https://www.heigit.org",
-        "description": "HEidelberg Institute for GeoInformation Technology"
+        "description": "Heidelberg Institute for GeoInformation Technology"
       }
     ],
     "fair:metrics_spec": [
