+++ models/yolo_swag_waste_grid_segmentation/tests/test_sampling.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from models.yolo11xcls_waste_grid_segmentation.pipeline import _subset_chips_dir
+from models.yolo_swag_waste_grid_segmentation.pipeline import _subset_chips_dir
 
 
 def _populate(chips_dir: Path, n: int) -> None:
